Move to gamestate namespace. Rename Gameplay to FreeRoam
