+++ Player/InteractablesManager.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using Godot;
 using untitledplantgame.Common;
+using untitledplantgame.Common.GameState;
 using untitledplantgame.Player;
 
 public partial class InteractablesManager : Node
@@ -14,7 +15,7 @@ public partial class InteractablesManager : Node
 	// In Player
 	public void ScanForInteractables()
 	{
-		if (GameStateMachine.Instance.CurrentState != GameState.Gameplay)
+		if (GameStateMachine.Instance.CurrentState != GameState.FreeRoam)
 		{
 			return;
 		}
