+++ project.godot
@@ -27,11 +27,11 @@ config/icon="res://icon.svg"
 
 [autoload]
 
-GameStateMachine="*res://Common/GameStateMachine.cs"
 EventBus="*res://EventBus/EventBus.cs"
 ResourceManager="*res://ResourceData/ResourceManager.cs"
 TimeController="*res://Common/TimeController.cs"
 WeatherCycle="*res://Cycle/Weather/WeatherCycle.cs"
+GameStateMachine="*res://Common/GameState/GameStateMachine.cs"
 
 [display]
